install.sh meshblu: guard 'docker-compose create'
'docker-compose create' fails if no "meshblu network" exists,
and does not attempt to create it although it is a dependency.

'docker-compose up' on the other hand does create the network
as specified in docker-compose.yml, but does more than just
installing the required components - it starts them as well.

no docker-compose command currently exists to auto-create the
network (alone) as specified in the docker-compose.yml file.